removing unecessary buffer constant value and checking type before unlinking tmpfile

+++ lib/audit_tentacles/media.rb
@@ -1,7 +1,5 @@
 class Media
   # Ensure we get tempfiles rather than stringio objects
-  OpenURI::Buffer.send :remove_const, 'StringMax' if OpenURI::Buffer.const_defined?('StringMax')
-  OpenURI::Buffer.const_set 'StringMax', 0
 
   def self.log(uri, context)
     unless $redis.sismember "#{$options.global_prefix}:uris", uri
@@ -26,7 +24,7 @@ class Media
     file = get_file(uri)
     info[:size] = file.size
     info[:sum] = hash_file(file)
-    file.unlink
+    file.unlink unless file.class == StringIO
     info
   end
